feat!: release v2.0.0 — address PR #50 review feedback
BREAKING CHANGE: bump to 2.0.0. ready() resolves true (not the Error),
unit IDs with astral characters now hash to canonical UTF-8, and
audienceMismatch cache invalidation changes for indeterminate audiences.

Review feedback from @calthejuggler:
- Revert gratuitous input-validation throws on context methods
  (attribute/peek/treatment/track/variableValue/peekVariableValue/
  override/customAssignment/customFieldValue/customFieldValueType and
  experiments() return); keep only correctness-motivated breaks.
- getAttribute(): reverse-scan early-return instead of full-array mutation.
- setInterval refresh scheduling collapsed to a one-liner.
- sdk.ts: iterate typed clientOptionKeys to drop the double cast; hoist
  context-option default constants to module scope.
- fetch.ts: single typed cast for the global fetch bind.
- Reorganize tests: remove fixes.test.js (ReDoS tests stale after revert,
  EqualsOperator dup'd into eq.test.js) and co-locate the rest into
  context/sdk/client/abort-controller-shim/fetch-shim test files; rename
  fixes-constructors.test.js -> constructor-options.test.js, deduped into
  one describe block.
- Rewrite README migration guide for v1 -> v2.

npm test 997/997, compile clean, lint clean.
